Move auth types to types file, add tsdoc for types
