+++ client/src/core/types.ts
@@ -7,3 +7,35 @@ export type OperationalTime = {
   hour: number;
   minute: number;
 }
+
+/** Represents the useAuth hook return object */
+export type AuthValue = {
+  /** @property authentication token stored in localStorage */
+  token?: string | null;
+
+  /** @property user data stored in localStorage  */
+  user?: UserData | null;
+
+  /** @property `true` if user is authenticating, `false` otherwise */
+  loading?: boolean;
+
+  /** @property method for logging in user with given Credentials */
+  loginAction: (credentials: Credentials) => Promise<void>;
+
+  /** @property method for logging out current user */
+  logoutAction: () => Promise<void>;
+};
+
+/** Represents the username and password credentials for logging in a user */
+export type Credentials = {
+  username: string,
+  password: string
+}
+
+/** Represent the returned user data after a user has logged in */
+export type UserData = {
+  username: string,
+  name: string,
+  surname: string
+  email: string
+};
